refactor(studio): split the global browser's page surface
GlobalDataBrowser was 362 lines, now 281.

- `GlobalDataPage` — everything that is a function of ONE loaded page: the
  drill-down chips, the grid, the pager. The browser keeps discovery, the poll,
  and the facet state.
- `GlobalTablesEmptyState` — the no-global-tables case, mostly to get eighteen
  lines of inline SVG out from between two conditionals.
- `global-row-format.ts` — row identity and chip values, shared by both and
  testable without a renderer.

`removeFilter` now takes an index instead of a MouseEvent it read `data-index`
back out of — the child already has the index, so round-tripping it through a
DOM attribute was the parent reaching into the child's markup.

Suppression deleted. 171 data tests green.

Refs #230.
